App now asks for location services to populate with current location
Refactored Some code from viewController
