+++ Fowl Weather/Fowl Weather/Helpers/ViewController+Extension+CoreLocation.swift
@@ -1,0 +1,48 @@
+//
+//  ViewController+Extension.swift
+//  Fowl Weather
+//
+//  Created by Jake Connerly on 11/20/19.
+//  Copyright © 2019 jake connerly. All rights reserved.
+//
+
+import Foundation
+import CoreLocation
+
+extension ViewController: CLLocationManagerDelegate {
+    
+    func checkLocationServices() {
+        if CLLocationManager.locationServicesEnabled() {
+            setupLocationManager()
+            checkLocationAuthorization()
+        } else {
+            //Pop alert
+        }
+    }
+    
+    func checkLocationAuthorization() {
+        switch CLLocationManager.authorizationStatus() {
+        case .authorizedWhenInUse:
+            locationManger.startUpdatingLocation()
+        case .denied:
+            break
+        case .notDetermined:
+            locationManger.requestWhenInUseAuthorization()
+        case .restricted:
+            break
+        case .authorizedAlways:
+            break
+        @unknown default:
+            fatalError()
+        }
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        userLocation = locations.last
+        performFetchByLocation(location: userLocation!)
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
+        checkLocationAuthorization()
+    }
+}
