+++ Fowl Weather/Fowl Weather/Resources/Info.plist
@@ -58,5 +58,7 @@
 		<string>UIInterfaceOrientationLandscapeLeft</string>
 		<string>UIInterfaceOrientationLandscapeRight</string>
 	</array>
+	<key>NSLocationWhenInUseUsageDescription</key>
+	<string>$(PRODUCT_NAME) is asking for your location to update your current weather.</string>
 </dict>
 </plist>
