+++ Fowl Weather/Fowl Weather/View Controllers/ViewController.swift
@@ -9,6 +9,7 @@
 import UIKit
 import ScalingCarousel
 import SHSearchBar
+import CoreLocation
 
 enum WeatherType: String {
     case clear = "clear sky"
@@ -55,6 +56,9 @@ class ViewController: UIViewController {
         }
     }
     
+    // Location Manager
+    let locationManger = CLLocationManager()
+    var userLocation: CLLocation?
     
     var sunriseDateFormatter: DateFormatter {
         let formatter = DateFormatter()
@@ -78,86 +82,10 @@ class ViewController: UIViewController {
         carouselCollectionView.dataSource = self
         carouselCollectionView.delegate = self
         observeSearchTerm()
+        checkLocationServices()
     }
-    
-    
     
     // MARK: - IBActions & Methods
-    
-    
-    @objc private func performFetches() {
-        
-        guard let searchTerm = searchTerm else { return }
-        var searchTermChecker: String?
-        let numString: [Character] = ["0","1","2","3","4","5","6","7","8","9"]
-        let searchByZip = Int(searchTerm)
-        for char in searchTerm {
-            for num in numString {
-                if char == num {
-                    searchTermChecker = String(char)
-                }
-            }
-        }
-        if searchByZip == nil && !searchTerm.isEmpty && searchTermChecker == nil {
-            performFetchByName(searchTerm: searchTerm)
-        } else if searchByZip != nil {
-            performFetchByZip(searchTerm: searchTerm)
-        } else if searchByZip == nil && searchTermChecker != nil {
-            popAlertControllerWithMessage(message: "\(searchTerm) is an invalid entry. Please try again.", title: "Invalid Search Entry")
-        } else {
-            //TODO: Pop alert Invalid entry
-        }
-        
-        jokeController.fetchRandomJoke { (_, error) in
-            if let error = error {
-                NSLog("Error fetching joke: \(error)")
-                return
-            }
-        }
-        
-        Group.dispatchGroup.notify(queue: .main) {
-            self.currentWeather = self.weatherController.currentWeather
-            if self.currentWeather == nil {
-                self.popAlertControllerWithMessage(message: "\(searchTerm) is an invalid entry. Please try again.", title: "Invalid Search Entry")
-                return
-            }
-            self.fiveDayForecast = self.weatherController.fiveDayForcast
-            self.joke = self.jokeController.joke
-            self.updateViews()
-            self.carouselCollectionView.reloadData()
-        }
-    }
-    
-    private func performFetchByZip(searchTerm: String) {
-        weatherController.fetchWeatherByZipCode(zipCode: searchTerm) { (weather, error) in
-            if let error = error {
-                NSLog("Error fetching current weather: \(error)")
-                return
-            }
-        }
-        
-        weatherController.fetchFiveDayByZipCode(zipCode: searchTerm) { (_, error) in
-            if let error = error {
-                NSLog("Error fetching current forecast: \(error)")
-                return
-            }
-        }
-    }
-    
-    private func performFetchByName(searchTerm: String) {
-        weatherController.fetchWeatherByCityName(cityName: searchTerm) { (_, error) in
-            if let error = error {
-                NSLog("Error fetching current weather: \(error)")
-                return
-            }
-        }
-        weatherController.fetchFiveDayByCityName(cityName: searchTerm) { (_, error) in
-            if let error = error {
-                NSLog("Error fetching current forecast: \(error)")
-                return
-            }
-        }
-    }
     
     private func observeSearchTerm() {
         NotificationCenter.default.addObserver(self, selector: #selector(didReceiveSearchTerm(_:)), name: .searchTermChosen, object: nil)
@@ -168,7 +96,7 @@ class ViewController: UIViewController {
         self.searchTerm = searchTerm
     }
     
-    private func updateViews() {
+    func updateViews() {
         guard let joke = joke,
             let currentWeather = currentWeather else {return}
         setBackground()
@@ -215,53 +143,9 @@ class ViewController: UIViewController {
         }
     }
     
-    private func cardinalDirectionHandler(directionInDegrees: Int) -> String {
-        var cardinalDirection: String = ""
-        switch directionInDegrees {
-        case (0...11):
-            cardinalDirection = "N"
-        case (12...33):
-            cardinalDirection = "N/NE"
-        case (34...56):
-            cardinalDirection = "NE"
-        case (57...78):
-            cardinalDirection = "E/NE"
-        case (79...101):
-            cardinalDirection = "E"
-        case (102...123):
-            cardinalDirection = "E/SE"
-        case (124...146):
-            cardinalDirection = "SE"
-        case (147...168):
-            cardinalDirection = "S/SE"
-        case (169...191):
-            cardinalDirection = "S"
-        case (192...213):
-            cardinalDirection = "S/SW"
-        case (214...236):
-            cardinalDirection = "SW"
-        case (237...258):
-            cardinalDirection = "W/SW"
-        case (259...281):
-            cardinalDirection = "W"
-        case (282...303):
-            cardinalDirection = "W/NW"
-        case (304...326):
-            cardinalDirection = "NW"
-        case (327...348):
-            cardinalDirection = "N/NW"
-        case (349...360):
-            cardinalDirection = "N"
-        default:
-            cardinalDirection = ""
-        }
-        return cardinalDirection
-    }
-    
-    private func popAlertControllerWithMessage(message: String, title: String) {
-        let alertController = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        alertController.addAction(UIAlertAction(title: "OK", style: .default, handler: nil))
-        self.present(alertController, animated: true, completion: nil)
+    func setupLocationManager() {
+        locationManger.delegate = self
+        locationManger.desiredAccuracy = kCLLocationAccuracyBest
     }
 }
